Open subjects with Subjects modal
