+++ src/components/Modals/Subjects/Subjects.js
@@ -99,7 +99,7 @@ export default function Subjects ({
             const char = i === subjectIndex ? `\u25CF` : `\u25CB`
             return (
               <Button
-                key={`SUBJECTS_${subj.id}`}
+                key={`SUBJECTS_PAGE_${i}`}
                 a11yTitle={`Go to subject page ${i}`}
                 onClick={() => changeSubjectIndex(i)}
               >
